Documentation
- Added `About` section
- Added `Stock-Data Suite Execution Example` as an execution guideline as requested in the assignment guidelines
- Added a `Known Bugs` section

+++ predictor.py
@@ -104,7 +104,7 @@ def queryTicker(fName, ticker, col):
 
     for row in info:
         if row["Ticker"] == ticker:
-            data.append((row["Time"], int(row[col])))
+            data.append((row["Time"], float(row[col])))
 
     if len(data) > 1:
         return data
